move functions into single file
