+++ detox/src/devices/android/ADB.js
@@ -5,19 +5,12 @@ const {execWithRetriesAndLogs, spawnAndLog} = require('../../utils/exec');
 const {escape} = require('../../utils/pipeCommands');
 const DetoxRuntimeError = require('../../errors/DetoxRuntimeError');
 const EmulatorTelnet = require('./EmulatorTelnet');
-const Environment = require('../../utils/environment');
+const {getAdbPath} = require('../../utils/environment');
 
 class ADB {
   constructor() {
     this._cachedApiLevels = new Map();
-    const platformToolsDir = path.join(Environment.getAndroidSDKPath(), 'platform-tools')
-    this.adbBin =
-      which.sync('adb', {path: platformToolsDir, nothrow: true}) ||
-      which.sync('adb', {nothrow: true});
-
-    if (this.adbBin == null){
-      throw new Error(Environment.MISSING_SDK_ERROR)
-    }
+    this.adbBin = getAdbPath();
   }
 
   async devices() {
